optimize folder and add static checker
