+++ libraries/botbuilder-lg/src/index.ts
@@ -1,8 +1,8 @@
 export * from './templateEngine';
 export * from './evaluator';
-export * from './LGFileLexer';
-export * from './LGFileParser';
-export * from './LGFileParserListener';
-export * from './LGFileParserVisitor';
+export * from './generator/LGFileLexer';
+export * from './generator/LGFileParser';
+export * from './generator/LGFileParserListener';
+export * from './generator/LGFileParserVisitor';
 export * from './getMethodExtensions';
 export * from './getValueExtensions';
